Feat: Refactoring the factory package for agroconnect

+++ src/main/java/com/agroconnect/frogger/factory/DefaultPaymentFactory.java
@@ -5,14 +5,16 @@ import com.agroconnect.frogger.entity.PaymentMethod;
 import com.agroconnect.frogger.entity.PaymentStatus;
 import org.springframework.stereotype.Component;
 
+import java.math.BigInteger;
+
 @Component
 public class DefaultPaymentFactory implements PaymentFactory {
 
     @Override
-    public Payment createPayment(Integer orderId, Integer customerId, double amount, String paymentMethod, String status) {
+    public Payment createPayment(BigInteger orderId, BigInteger customerId, double amount, String paymentMethod, String status) {
         Payment payment = new Payment();
-        payment.setOrderId(Long.valueOf(orderId));
-        payment.setCustomerId(Long.valueOf(customerId));
+        payment.setOrderId(orderId);
+        payment.setCustomerId(customerId);
         payment.setTotalAmount(amount);
         payment.setPaymentMethod(PaymentMethod.valueOf(paymentMethod));
         payment.setPaymentStatus(PaymentStatus.valueOf(status));
